+++ packages/@webex/plugin-meetings/src/index.js
@@ -13,4 +13,6 @@ export default Meetings;
 export * as CONSTANTS from './constants';
 export * as REACTIONS from './reactions/reactions';
 
+export {RemoteMedia} from './multistream/remoteMedia';
+
 export {default as TriggerProxy} from './common/events/trigger-proxy';